move RackApp into Servers folder and don't include in main library
