+++ config.ru
@@ -1,3 +1,7 @@
 require_relative 'lib/ical_filter_proxy'
+require_relative 'lib/ical_filter_proxy/servers/rack_app'
 
-run IcalFilterProxy.start_rack_app
+filters = IcalFilterProxy.filters
+app = IcalFilterProxy::Servers::RackApp.new(filters)
+
+run app
